requirements.txt
streamlit
openai
langchain
langchain-community
langchain-core
chromadb
python-dotenv
PyYAML
bs4
